Add support for adding external infoProcessors
A common use case is to create custom infoProcessors. Before you had to manually insert the processors in the module itself.

Fix this by adding an interface for dynamically adding new processors using the addInfoProcessor() function.
Furthermore, an interface is set for the processors, making it easier to create new ones. See BaseInfoProcessor class.