[vcfiterator] Add ability to ignore exceptions
In some cases (for instance when using malformed vcf files), it can still
be of interest to skip the failing line and continue processing.

Add ability to send in a throw_exceptions=False keyword to indicate that
exceptions should not throw. In that case, a warning is instead printed
to stderr.