[vcfiterator] Add include_raw functionality
Previously there was no way to get the raw data of the original
vcf file along with the parsed data.

Fix this by adding an include_raw parameter to make it return a
tuple, where one index is the raw line from the vcf. Useful for filtering
vcf files for instance.